feat(#5102): add Lint class for XMIR linting with WPA support
